delete unused ui
