+++ src/app/page.tsx
@@ -1,11 +1,9 @@
-import Navbar from "@/components/Navbar";
-import Tamagotchi from "@/components/Tamagotchi";
-import { TamagotchiNewUI } from "@/components/tamagotchi-new-ui";
+import { Tamagotchi } from "@/components/Tamagotchi";
 
 export default function Home() {
   return (
     <main className="flex min-h-screen flex-col items-center justify-center">
-      <TamagotchiNewUI />
+      <Tamagotchi />
     </main>
   );
 }
