+++ src/components/Tamagotchi.tsx
@@ -1,310 +1,359 @@
-// "use client";
-// import { INTERACTION } from "@/app/utils/interaction";
-// import React, { useState, useEffect } from "react";
-// import { idle, drinkingCoffee, death } from "./tamagotchiFrames";
+"use client";
+/**
+ * This code was generated by v0 by Vercel.
+ * @see https://v0.dev/t/JO1Z5J3ewhA
+ */
 
-// import "../index.css";
+import {
+  CardTitle,
+  CardDescription,
+  CardHeader,
+  CardContent,
+  CardFooter,
+  Card,
+} from "@/components/ui/card";
+import { Button } from "@/components/ui/button";
+import {
+  FaBowlFood,
+  FaBasketball,
+  FaBath,
+  FaFaceSadCry,
+  FaChartArea,
+} from "react-icons/fa6";
+import { useEffect, useState } from "react";
+import { death, idle } from "./tamagotchiFrames";
+import { MdLocalHospital } from "react-icons/md";
+import { INTERACTION } from "@/app/utils/interaction";
 
-// const DEFAULT_STATUS = ":)";
-// const Tamagotchi: React.FC = () => {
-//   const [frameIndex, setFrameIndex] = useState<number>(0);
-//   const [tamagotchiState, setTamagotchiState] = useState<any>({});
-//   const [animation, setAnimation] = useState<string[]>(idle);
-//   const [tamaStatus, setTamaStatus] = useState(DEFAULT_STATUS);
-//   const [checkingStatus, setCheckingStatus] = useState(false);
-//   const [isInteracting, setIsInteracting] = useState(false);
+const DEFAULT_STATUS = ":)";
+export function Tamagotchi() {
+  const [frameIndex, setFrameIndex] = useState<number>(0);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [loadingStatus, setLoadingStatus] = useState<string>("");
+  const [tamagotchiState, setTamagotchiState] = useState<any>({});
+  const [animation, setAnimation] = useState<string[]>(idle);
+  const [tamaStatus, setTamaStatus] = useState(DEFAULT_STATUS);
+  const [checkingStatus, setCheckingStatus] = useState(false);
+  const [isInteracting, setIsInteracting] = useState(false);
 
-//   //TODO - call init endpoint to determine if tamagotchi is initialized. if not generate one.
+  //TODO - call init endpoint to determine if tamagotchi is initialized. if not generate one.
 
-//   useEffect(() => {
-//     const fetchData = async () => {
-//       try {
-//         const response = await fetch("/api/getState", { method: "POST" });
-//         if (response.ok) {
-//           const jsonData = await response.json();
-//           if (!!!jsonData.comment) {
-//             jsonData.comment = "No comments";
-//           }
-//           setTamagotchiState(jsonData);
-//           console.log(jsonData);
-//           handleDeath(jsonData, pollInterval);
-//         }
-//       } catch (error) {
-//         console.error("Error fetching data:", error);
-//       }
-//     };
+  useEffect(() => {
+    let progress = 0;
+    let loadingInterval: NodeJS.Timer;
+    if (isLoading) {
+      loadingInterval = setInterval(() => {
+        const loadingStr = ".".repeat(progress % 5);
+        setLoadingStatus(loadingStr);
+        progress++;
+      }, 200);
+    } else {
+      setLoadingStatus("");
+    }
 
-//     // Cycle through frames every 1 second
-//     const frameInterval = setInterval(() => {
-//       setFrameIndex((prevIndex) => (prevIndex + 1) % idle.length);
-//     }, 1000);
+    const fetchData = async () => {
+      try {
+        const response = await fetch("/api/getState", { method: "POST" });
+        if (response.ok) {
+          const jsonData = await response.json();
+          if (!!!jsonData.comment) {
+            jsonData.comment = "No comments";
+          }
+          setTamagotchiState(jsonData);
+          console.log(jsonData);
+          handleDeath(jsonData, pollInterval);
+        }
+      } catch (error) {
+        console.error("Error fetching data:", error);
+      }
+    };
 
-//     // Fetch data on component mount
-//     fetchData();
+    // Cycle through frames every 1 second
+    const frameInterval = setInterval(() => {
+      setFrameIndex((prevIndex) => (prevIndex + 1) % idle.length);
+    }, 500);
 
-//     // Start polling data every N seconds (adjust the interval as needed)
-//     const pollInterval = setInterval(fetchData, 5000); // Poll every 5 seconds
+    // Fetch data on component mount
+    fetchData();
 
-//     return () => {
-//       clearInterval(pollInterval);
-//       clearInterval(frameInterval);
-//     };
-//   }, []);
+    // Start polling data every N seconds (adjust the interval as needed)
+    const pollInterval = setInterval(fetchData, 5000); // Poll every 5 seconds
 
-//   const handleResponse = (responseText: string) => {
-//     const responseJSON = JSON.parse(responseText);
-//     const animation = JSON.parse(responseJSON.animation);
-//     const status = responseJSON.status;
-//     setFrameIndex(0);
-//     setAnimation(animation);
-//     setTamaStatus(status);
-//   };
+    return () => {
+      clearInterval(pollInterval);
+      clearInterval(frameInterval);
+      clearInterval(loadingInterval);
+    };
+  }, [isLoading]);
 
-//   const handleDeath = (jsonData: any, pollInterval: NodeJS.Timer) => {
-//     if (jsonData.death) {
-//       setAnimation(death);
-//       setTamaStatus("Dead :(");
-//       clearInterval(pollInterval);
-//     }
-//   };
+  const handleResponse = (responseText: string) => {
+    const responseJSON = JSON.parse(responseText);
+    const animation = JSON.parse(responseJSON.animation);
+    const status = responseJSON.status;
+    setFrameIndex(0);
+    setAnimation(animation);
+    setTamaStatus(status);
+  };
 
-//   const handleBath = async () => {
-//     setIsInteracting(true);
-//     setTamaStatus("Bathing");
-//     try {
-//       const response = await fetch("/api/interact", {
-//         method: "POST",
-//         body: JSON.stringify({
-//           interactionType: INTERACTION.BATH,
-//         }),
-//         headers: {
-//           "Content-Type": "application/json",
-//         },
-//       });
-//       const responseText = await response.text();
-//       handleResponse(responseText);
-//     } catch (e) {
-//       console.log(e);
-//     }
+  const handleDeath = (jsonData: any, pollInterval: NodeJS.Timer) => {
+    //TODO - not yet implemented
+    if (jsonData.death) {
+      setAnimation(death);
+      setTamaStatus("Dead :(");
+      clearInterval(pollInterval);
+    }
+  };
 
-//     setTimeout(() => {
-//       setAnimation(idle);
-//       setTamaStatus(DEFAULT_STATUS);
-//       setIsInteracting(false);
-//     }, 9000);
-//   };
+  const handleBath = async () => {
+    setIsLoading(true);
+    setIsInteracting(true);
+    setTamaStatus("Bathing");
+    try {
+      const response = await fetch("/api/interact", {
+        method: "POST",
+        body: JSON.stringify({
+          interactionType: INTERACTION.BATH,
+        }),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      const responseText = await response.text();
+      handleResponse(responseText);
+    } catch (e) {
+      console.log(e);
+    }
 
-//   const feedTamagotchi = async (e: any) => {
-//     setIsInteracting(true);
-//     // Add logic to feed the Tamagotchi here
-//     setTamaStatus("Feeding...");
-//     try {
-//       const response = await fetch("/api/interact", {
-//         method: "POST",
-//         body: JSON.stringify({
-//           interactionType: INTERACTION.FEED,
-//         }),
-//         headers: {
-//           "Content-Type": "application/json",
-//         },
-//       });
-//       const responseText = await response.text();
-//       handleResponse(responseText);
-//     } catch (e) {
-//       console.log(e);
-//     }
+    setTimeout(() => {
+      setAnimation(idle);
+      setTamaStatus(DEFAULT_STATUS);
+      setIsInteracting(false);
+      setIsLoading(false);
+    }, 9000);
+  };
 
-//     console.log("Tamagotchi fed!");
+  const feedTamagotchi = async (e: any) => {
+    setIsInteracting(true);
+    setIsLoading(true);
+    // Add logic to feed the Tamagotchi here
+    setTamaStatus("Feeding");
+    try {
+      const response = await fetch("/api/interact", {
+        method: "POST",
+        body: JSON.stringify({
+          interactionType: INTERACTION.FEED,
+        }),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      const responseText = await response.text();
+      setIsLoading(false);
+      handleResponse(responseText);
+    } catch (e) {
+      console.log(e);
+    }
 
-//     // TODO - there may be a race condition here if user clicks the button too fast?
-//     setTimeout(() => {
-//       setAnimation(idle);
-//       setTamaStatus(DEFAULT_STATUS);
-//       setIsInteracting(false);
-//     }, 9000);
-//   };
+    console.log("Tamagotchi fed!");
 
-//   const playWithTamagotchi = async (e: any) => {
-//     // Add logic to feed the Tamagotchi here
-//     setTamaStatus("Playing...");
-//     setIsInteracting(true);
-//     try {
-//       const response = await fetch("/api/interact", {
-//         method: "POST",
-//         body: JSON.stringify({
-//           interactionType: INTERACTION.PLAY,
-//         }),
-//         headers: {
-//           "Content-Type": "application/json",
-//         },
-//       });
-//       const responseText = await response.text();
-//       handleResponse(responseText);
-//     } catch (e) {
-//       console.log(e);
-//     }
-//     setTimeout(() => {
-//       setAnimation(idle);
-//       setTamaStatus(DEFAULT_STATUS);
-//       setIsInteracting(false);
-//     }, 9000);
-//   };
+    // TODO - there may be a race condition here if user clicks the button too fast?
+    setTimeout(() => {
+      setAnimation(idle);
+      setTamaStatus(DEFAULT_STATUS);
+      setIsInteracting(false);
+    }, 9000);
+  };
 
-//   const treatSickTamagotchi = async (e: any) => {
-//     setIsInteracting(true);
-//     try {
-//       const response = await fetch("/api/interact", {
-//         method: "POST",
-//         body: JSON.stringify({
-//           interactionType: INTERACTION.GO_TO_HOSPITAL,
-//         }),
-//         headers: {
-//           "Content-Type": "application/json",
-//         },
-//       });
-//       const responseText = await response.text();
-//       handleResponse(responseText);
-//     } catch (e) {
-//       console.log(e);
-//     }
-//     setTimeout(() => {
-//       setAnimation(idle);
-//       setTamaStatus(DEFAULT_STATUS);
-//       setIsInteracting(false);
-//     }, 9000);
-//   };
+  const playWithTamagotchi = async (e: any) => {
+    // Add logic to feed the Tamagotchi here
+    setIsLoading(true);
+    setTamaStatus("Playing");
+    setIsInteracting(true);
+    try {
+      const response = await fetch("/api/interact", {
+        method: "POST",
+        body: JSON.stringify({
+          interactionType: INTERACTION.PLAY,
+        }),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      const responseText = await response.text();
+      setIsLoading(false);
+      handleResponse(responseText);
+    } catch (e) {
+      console.log(e);
+    }
+    setTimeout(() => {
+      setAnimation(idle);
+      setTamaStatus(DEFAULT_STATUS);
+      setIsInteracting(false);
+    }, 9000);
+  };
 
-//   const handleDiscipline = async () => {
-//     setTamaStatus("Discipling...");
-//     setIsInteracting(true);
-//     try {
-//       const response = await fetch("/api/interact", {
-//         method: "POST",
-//         body: JSON.stringify({
-//           interactionType: INTERACTION.DISCIPLINE,
-//         }),
-//         headers: {
-//           "Content-Type": "application/json",
-//         },
-//       });
-//       const responseText = await response.text();
-//       handleResponse(responseText);
-//     } catch (e) {
-//       console.log(e);
-//     }
-//     setTimeout(() => {
-//       setAnimation(idle);
-//       setTamaStatus(DEFAULT_STATUS);
-//       setIsInteracting(false);
-//     }, 9000);
-//   };
+  const treatSickTamagotchi = async (e: any) => {
+    setIsLoading(true);
+    setTamaStatus("Going to hospital");
+    setIsInteracting(true);
+    try {
+      const response = await fetch("/api/interact", {
+        method: "POST",
+        body: JSON.stringify({
+          interactionType: INTERACTION.GO_TO_HOSPITAL,
+        }),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      const responseText = await response.text();
+      handleResponse(responseText);
+    } catch (e) {
+      console.log(e);
+    }
+    setTimeout(() => {
+      setAnimation(idle);
+      setTamaStatus(DEFAULT_STATUS);
+      setIsInteracting(false);
+      setIsLoading(false);
+    }, 9000);
+  };
 
-//   const checkStatus = () => {
-//     if (!isInteracting) {
-//       setTamaStatus("Checking Status...");
-//       setCheckingStatus(true);
-//       setTimeout(() => {
-//         setCheckingStatus(false);
-//       }, 9000);
-//     }
-//   };
+  const handleDiscipline = async () => {
+    setIsLoading(true);
+    setTamaStatus("Discipling");
+    setIsInteracting(true);
+    try {
+      const response = await fetch("/api/interact", {
+        method: "POST",
+        body: JSON.stringify({
+          interactionType: INTERACTION.DISCIPLINE,
+        }),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      const responseText = await response.text();
+      setIsLoading(false);
+      handleResponse(responseText);
+    } catch (e) {
+      console.log(e);
+    }
+    setTimeout(() => {
+      setAnimation(idle);
+      setTamaStatus(DEFAULT_STATUS);
+      setIsInteracting(false);
+    }, 9000);
+  };
 
-//   let needForBath = "";
-//   for (let i = 0; i < tamagotchiState.poop; i++) {
-//     needForBath += "💩";
-//   }
+  const checkStatus = () => {
+    setIsLoading(true);
+    if (!isInteracting) {
+      setTamaStatus("Checking Status");
+      setCheckingStatus(true);
+      setTimeout(() => {
+        setAnimation(idle);
+        setTamaStatus(DEFAULT_STATUS);
+        setIsInteracting(false);
+        setCheckingStatus(false);
+        setIsLoading(false);
+      }, 6000);
+    }
+  };
 
-//   return (
-//     <div className="flex flex-col justify-center items-center h-screen w-screen bg-slate-50 tamago-frame">
-//       <div className="text-center mb-2">Status: {tamaStatus}</div>
-//       <div className="min-w-[200px] p-4 border border-grey rounded-lg  h-[230px] overflow-auto tamago-screen">
-//         {/* Tamagotchi display */}
-//         <div className="flex justify-center">
-//           <div className="flex items-center justify-center w-64 h-48 overflow-hidden">
-//             {!checkingStatus && (
-//               <div>
-//                 <pre className="text-center">{animation[frameIndex]}</pre>
-//                 <div>{needForBath}</div>
-//               </div>
-//             )}
+  let needForBath = "";
+  for (let i = 0; i < tamagotchiState.poop; i++) {
+    needForBath += "💩";
+  }
+  return (
+    <Card key="1" className="w-full max-w-xl p-6 space-y-6 bg-white">
+      <CardHeader className="items-center space-y-2">
+        <CardTitle className="text-xl">AI Tamagotchi</CardTitle>
+        <CardDescription className="text-sm text-gray-500 dark:text-gray-400">
+          Care for your virtual pet!
+        </CardDescription>
+        <div className="flex justify-center">
+          <div className="flex items-center justify-center h-48 overflow-hidden">
+            {!checkingStatus && (
+              <div>
+                <pre className="text-center">{animation[frameIndex]}</pre>
+                <div>{needForBath}</div>
+              </div>
+            )}
 
-//             {checkingStatus && (
-//               <pre
-//                 className="text-left overflow-x-hidden overflow-y-auto whitespace-normal"
-//                 style={{ maxHeight: "100%", maxWidth: "100%" }}
-//               >
-//                 Age: {tamagotchiState.age || "No age"} <br />
-//                 Happiness: {tamagotchiState.happiness || "No happiness"} <br />
-//                 Hunger: {tamagotchiState.hunger || "No hunger"} <br />
-//                 Health: {tamagotchiState.health || "No health"} <br />
-//                 🚽: {tamagotchiState.poop || "👍"} <br />
-//                 {'"' + tamagotchiState.comment || "No comments" + '"'}
-//               </pre>
-//             )}
-//           </div>
-//         </div>
-//       </div>
-//       <div className="flex justify-between mb-4">
-//         <button
-//           onClick={feedTamagotchi}
-//           className="px-4 py-2 mr-2 bg-blue-200 rounded-lg"
-//           style={{ width: "80px" }}
-//         >
-//           🍴
-//         </button>
-//         {/* TODO - all other actions */}
-//         {/* <button
-//             onClick={feedTamagotchi}
-//             className="px-4 py-2 mr-2 bg-blue-200 rounded-lg"
-//           >
-//             💡
-//           </button> */}
-//         <button
-//           onClick={playWithTamagotchi}
-//           className="px-4 py-2 mr-2 bg-blue-200 rounded-lg"
-//           style={{ width: "80px" }}
-//         >
-//           🎯
-//         </button>
-//         <button
-//           onClick={treatSickTamagotchi}
-//           className="px-4 py-2 mr-2 bg-blue-200 rounded-lg"
-//           style={{ width: "80px" }}
-//         >
-//           💉
-//         </button>
-//       </div>
-//       <div className="flex justify-between">
-//         <button
-//           onClick={handleBath}
-//           className="px-4 py-2 mr-2 bg-blue-200 rounded-lg"
-//           style={{ width: "80px" }}
-//         >
-//           🛀
-//         </button>
-//         <button
-//           onClick={checkStatus}
-//           className="px-4 py-2 mr-2 bg-blue-200 rounded-lg"
-//           style={{ width: "80px" }}
-//         >
-//           📋
-//         </button>
-//         <button
-//           onClick={handleDiscipline}
-//           className="px-4 py-2 mr-2 bg-blue-200 rounded-lg"
-//           style={{ width: "80px" }}
-//         >
-//           😡
-//         </button>
-//         {/* <button
-//             onClick={feedTamagotchi}
-//             className="px-4 py-2 mr-2 bg-blue-200 rounded-lg"
-//           >
-//             👥
-//           </button> */}
-//       </div>
-//     </div>
-//   );
-// };
-
-// export default Tamagotchi;
+            {checkingStatus && (
+              <pre
+                className="text-left overflow-x-hidden overflow-y-auto whitespace-normal"
+                style={{ maxHeight: "100%", maxWidth: "100%" }}
+              >
+                Age: {tamagotchiState.age || "No age"} <br />
+                Happiness: {tamagotchiState.happiness || "No happiness"} <br />
+                Hunger: {tamagotchiState.hunger || "No hunger"} <br />
+                Health: {tamagotchiState.health || "No health"} <br />
+                🚽: {tamagotchiState.poop || "👍"} <br />
+                {'"' + tamagotchiState.comment || "No comments" + '"'}
+              </pre>
+            )}
+          </div>
+        </div>
+      </CardHeader>
+      <CardContent className="grid grid-cols-2 gap-4">
+        <Button
+          onClick={feedTamagotchi}
+          className="flex items-center justify-center"
+          variant="outline"
+          disabled={isInteracting}
+        >
+          <FaBowlFood /> &nbsp; Feed
+        </Button>
+        <Button
+          onClick={playWithTamagotchi}
+          className="flex items-center justify-center"
+          variant="outline"
+          disabled={isInteracting}
+        >
+          <FaBasketball /> &nbsp; Play
+        </Button>
+        <Button
+          disabled={isInteracting}
+          onClick={handleBath}
+          className="flex items-center justify-center"
+          variant="outline"
+        >
+          <FaBath />
+          &nbsp; Clean
+        </Button>
+        <Button
+          disabled={isInteracting}
+          onClick={handleDiscipline}
+          className="flex items-center justify-center"
+          variant="outline"
+        >
+          <FaFaceSadCry /> &nbsp; Discipline
+        </Button>
+        <Button
+          disabled={isInteracting}
+          onClick={checkStatus}
+          className="flex items-center justify-center"
+          variant="outline"
+        >
+          <FaChartArea /> &nbsp; Status
+        </Button>
+        <Button
+          disabled={isInteracting}
+          onClick={treatSickTamagotchi}
+          className="flex items-center justify-center"
+          variant="outline"
+        >
+          <MdLocalHospital /> &nbsp; Medical Care
+        </Button>
+      </CardContent>
+      <CardFooter className="py-4 dark:bg-gray-800 rounded-b-lg border-t-2 border-gray-200 dark:border-gray-700">
+        <div className="h-10 text-center text-lg font-semibold text-gray-600 dark:text-gray-400">
+          {tamaStatus}
+          {loadingStatus}
+        </div>
+      </CardFooter>
+    </Card>
+  );
+}
